refactor(skeleton): update import path
